refactor: moving helpers and constants into util folder
